Move ActionSelection, BetaDistribution, MixtureModel, ThompsonSampling
to the parent folder since it is not specific to the backward chainer
