+++ opencog/rule-engine/BetaDistribution.h
