+++ opencog/rule-engine/ThompsonSampling.cc
