+++ opencog/rule-engine/ThompsonSampling.h
