+++ opencog/rule-engine/backwardchainer/CMakeLists.txt
@@ -2,10 +2,6 @@ INSTALL (FILES
 	BackwardChainer.h
 	TraceRecorder.h
 	ControlPolicy.h
-	MixtureModel.h
-	ActionSelection.h
-	BetaDistribution.h
-        ThompsonSampling.h
 	BIT.h
 	Fitness.h
 	DESTINATION "include/opencog/rule-engine/backwardchainer"
